alert_generation/socketio_server/requirements.txt
python-socketio
pyzmq
eventlet
aiohttp
